Improved EnumPrefix to use powers instad of really long numbers and added comments for missing prefixes.
